// FoodDiary.Web.Client/src/app/shared/api/sdk/generated/model/favorite-meal-http-response.ts
/**
 * FoodDiary User API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export interface FavoriteMealHttpResponse {
    createdAtUtc?: string;
    id?: string;
    imageUrl?: string | null;
    itemCount?: number;
    itemImageUrls?: Array<string>;
    itemNames?: Array<string>;
    mealDate?: string;
    mealId?: string;
    mealType?: string | null;
    name?: string | null;
    totalCalories?: number;
    totalCarbs?: number;
    totalFats?: number;
    totalFiber?: number;
    totalProteins?: number;
}