// FoodDiary.Web.Client/src/app/shared/api/sdk/generated/model/favorite-recipe-http-response.ts
/**
 * FoodDiary User API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export interface FavoriteRecipeHttpResponse {
    createdAtUtc?: string;
    id?: string;
    imageUrl?: string | null;
    ingredientCount?: number;
    ingredientNames?: Array<string>;
    name?: string | null;
    recipeId?: string;
    recipeName?: string;
    servings?: number;
    totalCalories?: number | null;
    totalCarbs?: number;
    totalFats?: number;
    totalFiber?: number;
    totalProteins?: number;
    totalTimeMinutes?: number | null;
}